Add support for extensions in log entries (#353)
